requirements.txt
psycopg2-binary==2.7.5
facebook-sdk
Pillow==4.2.1
django==1.11